+++ pkg/infra/httpserver/http.go
@@ -1,4 +1,4 @@
-package server
+package httpserver
 
 import (
 	"context"
