+++ pkg/word/model.go
